requirements.txt
asgiref==3.6.0
Django==4.2.1
django-filter==23.2
django-phonenumber-field==7.1.0
djangorestframework==3.14.0
phonenumberslite==8.13.11
Pillow==9.5.0
python-dotenv==1.0.0
pytz==2023.3
sqlparse==0.4.4
tzdata==2023.3
